% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nationalpark.R
\docType{package}
\name{nationalpark}
\alias{nationalpark}
\alias{nationalpark-package}
\title{nationalpark}
\description{
list of palettes from US National Park photographs
}
\details{
list of palettes from US National Park photographs
}
