% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{np_palette}
\alias{np_palette}
\title{A National Parks of the United States palette generator}
\usage{
np_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{badlands}, \code{deathvalley}, \code{smokies},
\code{picturedrocks}, \code{lakesuperior}, \code{tallgrass},
\code{rockymtn}, \code{flatirons}, \code{shenandoah},
\code{shenandoah2}, \code{kingscanyon}, \code{kingscanyon2},
\code{smokies2}, \code{smokies3}, \code{arches},
\code{bryce}, \code{grandtetons}, \code{conagree},
\code{zion}, \code{yellowstone}}

\item{n}{Number of colors desired. Unfortunately most palettes now only
have 5 colors. These palettes are picked using color.adobe.com
from photographs provided by J.W. Atkins, E. Agee, A. R. Woleslagle.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes pulled from photographs of US National Parks.
}
\examples{
np_palette("badlands")
np_palette("deathvalley")
np_palette("deathvalley", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- np_palette(21, name = "badlands", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
